# feat: Add additional EC key validation for FIPS