### Description of changes: 

The [EC_KEY_check_key](https://www.openssl.org/docs/manmaster/man3/EC_KEY_check_key.html) libcrypto API is used in s2n-tls to validate the public key received from the peer. Some libcryptos, such as AWS-LC, provide another API: [EC_KEY_check_fips](https://github.com/aws/aws-lc/blob/6d9a7675d1f7c2f159b76e4ade309220d54d96f4/include/openssl/ec_key.h#L171-L177), which includes the validation performed by `EC_KEY_check_key` as well as additional validation for FIPS.

This PR calls the `EC_KEY_check_fips` API instead of `EC_KEY_check_key` when s2n-tls is operating in FIPS mode.

### Call-outs:

`EC_KEY_check_fips` isn't defined in all libcryptos, so a new feature probe was added.

### Testing:

A new test was added to ensure that the feature probe is working correctly.

### Performance:

I ran benchmarks to determine the performance impact of this change with AWS-LC-FIPS-2022. I used [this s2n_perf unit test](https://github.com/aws/s2n-tls/commit/7556dcac9949fb2320d7cd7b7a4b992f527f57ad) to get results. I ran each operation for 10 minutes.

While this change did significantly impact the performance of `s2n_ecc_evp_compute_shared_secret_from_params`, this is such a small part of the handshake that there wasn't really a noticeable change.

operation | op time before change (us) | op time after change (us) | % slower
-- | -- | -- | --
s2n_ecc_evp_compute_shared_secret_from_params | 69.14 | 156.76 | 126.72%
TLS 1.3 handshake with secp256 | 1,461.52 | 1,464.08 | 0.17%

generate vs compute_shared_secret comparison:

operation | op time before change (us) | op time after change (us)
-- | -- | --
s2n_ecc_evp_generate_ephemeral_key | 108.37 | 108.37
s2n_ecc_evp_compute_shared_secret_from_params | 69.14 | 156.76
% of time checking vs generating | 38.95% | 59.12%

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.